changed timepuncher exceptions to time log
