+++ app/DomainValueObjects/TimeLog/TimeStamp/TimeStamp.php
@@ -5,7 +5,7 @@ namespace App\DomainValueObjects\TimeLog\TimeStamp;
 
 use App\DomainValueObjects\UUIDGenerator\UUID;
 use App\DomainValueObjects\Location\Location;
-use App\Exceptions\TimePuncherExceptions\TimeStamp\GenerateTimeStampFailed;
+use App\Exceptions\TimeLogExceptions\TimeStamp\GenerateTimeStampFailed;
 
 
 class TimeStamp
