+++ app/Http/Controllers/Api/TimeLog/ClockOutDomain/Services/ClockOutService.php
@@ -10,7 +10,7 @@ use Illuminate\Support\Facades\Auth;
 use App\TimeLog;
 
 //Exceptions 
-use App\Exceptions\TimePuncherExceptions\ClockOut\AlreadyClockedOut;
+use App\Exceptions\TimeLogExceptions\ClockOut\AlreadyClockedOut;
 // Contracts 
 use App\Http\Controllers\Api\TimeLog\ClockOutDomain\Contracts\ClockOutContract;
 use App\Http\Controllers\Api\TimeLog\Logic\Contracts\ClockOutLogicContract;
